+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Versions follows PEP440 version scheme.
 ************
 
 ## Fixed
+* fix: [RGG-936] Interval filter breaks after clicking "CLEAR" on empty field
 * fix: [RGG-908] Fix for custom selects in active state
 
 gamma-core 3.4.0 (2023-06-08)